functions/requirements.txt
# Azure Functions dependencies
azure-functions>=1.18.0
